+++ src/features/form/components/QuestionRenderers/RankingRenderer.tsx
@@ -1,6 +1,6 @@
 import { DragToOrder } from "@/shared/components";
 import type { FormsQuestionResponse } from "@nycu-sdc/core-system-sdk";
-import styles from "../FormDetailPage.module.css";
+import styles from "../FormFilloutPage.module.css";
 
 export const RankingRenderer = ({
 	question,
